backend/requirements.txt
fastapi
uvicorn
sqlmodel
pymssql
python-dotenv
openpyxl
python-jose[cryptography]
passlib[bcrypt]
httpx<0.28
pytest
python-multipart
a2wsgi
